contact page
to do:

text area scrollbar, testing, cleanup, check text
